Extract labby-primitives leaf crate; drop labby-apis from gateway/auth/codemode
labby-gateway, labby-auth, and labby-codemode only ever needed a handful of
small, self-contained types from labby-apis (ActionSpec/ParamSpec,
PluginMeta/EnvVar/Category, UiSchema, and the static SSRF checks) but paid
for the whole SDK crate (HTTP client, SSH helpers, 9 product service
modules) as a transitive dependency.

- New labby-primitives crate: those types moved verbatim, zero internal
  dependencies. labby-apis re-exports them for backward compatibility.
- labby-gateway depends on labby-primitives instead of labby-apis (labby-apis
  now dev-dependency only, for one test fixture).
- dispatch_helpers.rs and the spawn-guard/SSRF security guards moved from
  labby-runtime into labby-gateway — they were gateway-only concerns that
  pulled labby-primitives into labby-auth and labby-codemode's dependency
  graph for no reason (neither ever called into them).
- labby-runtime's labby-apis dependency is now optional, gated behind the
  marketplace/acp_registry/deploy features (the only remaining callers).
- labby-auth and labby-codemode now depend on exactly one other crate:
  labby-runtime.
- crates/labby's marketplace feature now explicitly requires gateway, since
  marketplace's install-time validation reuses the gateway-owned spawn-guard.
