+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "3"
-members = ["crates/labby-apis", "crates/labby", "crates/labby-auth", "crates/labby-codemode", "crates/labby-gateway", "crates/labby-web", "crates/labby-runtime", "crates/labby-winjob"]
+members = ["crates/labby-apis", "crates/labby", "crates/labby-auth", "crates/labby-codemode", "crates/labby-gateway", "crates/labby-web", "crates/labby-runtime", "crates/labby-winjob", "crates/labby-primitives"]
 
 [workspace.package]
 version = "0.29.0"
