+++ crates/labby-gateway/src/lib.rs
@@ -15,6 +15,7 @@
 //! The pool is re-exported from `lab`'s `crate::dispatch::upstream` as a
 //! compatibility shim so existing Labby callers keep working unchanged.
 
+pub mod dispatch_helpers;
 pub mod gateway;
 pub mod net;
 pub mod process;
