+++ crates/labby-gateway/src/security.rs
@@ -1,3 +1,7 @@
 //! Shared security preflight guards used by the gateway runtime.
+//!
+//! Owned here (not in `labby-runtime`) because both modules are gateway-only:
+//! neither `labby-auth` nor `labby-codemode` ever call into them.
 
-pub use labby_runtime::security::ssrf;
+pub mod spawn_guard;
+pub mod ssrf;
