+++ crates/labby-runtime/src/lib.rs
@@ -7,15 +7,18 @@
 //! contracts used by more than one extracted crate. It must not depend on
 //! product/transport layers (`axum`, `clap`, `rmcp`, `javy`, `wasmtime`,
 //! `utoipa`) or on Labby product registry builders.
+//!
+//! `dispatch_helpers` and the stdio-spawn/SSRF security guards moved to
+//! `labby-gateway` — they're gateway-only concerns, and keeping them here
+//! would pull the dependency-free `labby-primitives` types they use into
+//! `labby-auth` and `labby-codemode`, which never touch them.
 
 pub mod backoff;
-pub mod dispatch_helpers;
 pub mod error;
 pub mod gateway_config;
 pub mod helpers;
 pub mod path_safety;
 pub mod redact;
-pub mod security;
 
 pub use helpers::{env_non_empty, home_dir, lab_home};
