--- crates/labby-runtime/src/security.rs
@@ -1,4 +1,0 @@
-//! Surface-neutral security preflight guards shared by product runtime crates.
-
-pub mod spawn_guard;
-pub mod ssrf;
