+++ crates/labby/src/dispatch/marketplace/mcp_dispatch.rs
@@ -591,7 +591,7 @@ async fn mcp_client_config(
     prefs: &GatewayPreferences,
 ) -> Result<Value, ToolError> {
     if let Some(url) = server.remotes.iter().find_map(|r| r.url.as_deref()) {
-        labby_runtime::security::ssrf::validate_external_https_url(url).await?;
+        labby_gateway::security::ssrf::validate_external_https_url(url).await?;
         return Ok(serde_json::json!({ "url": url }));
     }
 
@@ -692,7 +692,7 @@ async fn install_http(
 ) -> Result<Value, ToolError> {
     let url = url.to_string();
 
-    labby_runtime::security::ssrf::validate_external_https_url(&url).await?;
+    labby_gateway::security::ssrf::validate_external_https_url(&url).await?;
 
     // Probe for OAuth support — non-fatal, install proceeds without OAuth on failure.
     #[cfg(feature = "gateway")]
